Fixes DCHECK crash when closing Tor window.
When Tor window is closed (via X button), the tor process is terminated
and tor launcher service tries to report the process crash. However the
browser side already reset the Remote in BrowserContextShutdown, so the
crash callback causes a DCHECK.

Added disconnect handler in tor_launcher_service that flips a boolean
which is checked when attempting to use the crash callback.

Also, modified TorProfileServiceFactory::BrowserContextShutdown to not
create TorProfileService for non-tor profiles.